+++ st_df_table/frontend/src/StTable.tsx
@@ -1,6 +1,8 @@
-import React, {useEffect} from "react";
+import React, {useEffect, useRef} from "react";
 import {Streamlit, withStreamlitConnection} from "streamlit-component-lib";
-import BootstrapTable, {ColumnDescription} from 'react-bootstrap-table-next';
+import BootstrapTable, {ColumnDescription, RowEventHandlerProps} from 'react-bootstrap-table-next';
+import paginationFactory from 'react-bootstrap-table2-paginator';
+
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'react-bootstrap-table-next/dist/react-bootstrap-table2.min.css';
 
@@ -9,28 +11,46 @@ interface Props {
 }
 
 const StTable: React.FC<Props> = (props) => {
+    const pagination_line_height = 10;
+    const padding = 8;
+    const pagination_extra_table_height = 54;
+
     const {args} = props;
-    const head_align = args.head_align
-    const data_align = args.data_align
-    const head_bg_color = args.head_bg_color
-    const data_bg_color = args.data_bg_color
-    const head_color = args.head_color
-    const data_color = args.data_color
-    const head_font_weight = args.head_font_weight
-    const data_font_weight = args.data_font_weight
-    const bordered = args.bordered
-    const border_color = args.border_color
-    let border_width = args.border_width
-    const table_width = args.table_width
-    const sortable = args.sortable
-    const font = args.font
-    const font_size = args.font_size
+    const {
+        head_align,
+        data_align,
+        head_bg_color,
+        data_bg_color,
+        head_color,
+        data_color,
+        head_font_weight,
+        data_font_weight,
+        bordered,
+        border_color,
+        border_width,
+        table_width,
+        sortable,
+        font,
+        font_size,
+        paginated,
+        pagination_size_per_page,
+        pagination_bar_size,
+        pagination_text_color,
+        pagination_bg_color,
+        pagination_border_color,
+        pagination_active_color,
+        pagination_active_border_color,
+        pagination_active_bg_color,
+        pagination_hover_color,
+        pagination_hover_bg_color,
+        vertical_alignment,
+        data,
+    } = args;
 
-    if (!bordered) {
-        border_width = 0
-    }
+    const tableRef = useRef<HTMLDivElement | null>(null);
+    console.log(vertical_alignment)
 
-    const columns = args.columns.map((col: ColumnDescription) => ({
+    const columns: ColumnDescription[] = args.columns.map((col: ColumnDescription) => ({
         ...col,
         sort: sortable,
         headerStyle: {
@@ -40,6 +60,7 @@ const StTable: React.FC<Props> = (props) => {
             fontWeight: head_font_weight,
             fontFamily: font,
             fontSize: font_size,
+            verticalAlign: vertical_alignment,
         },
         style: {
             textAlign: col.align || data_align,
@@ -48,40 +69,114 @@ const StTable: React.FC<Props> = (props) => {
             fontWeight: data_font_weight,
             fontFamily: font,
             fontSize: font_size,
+            whiteSpace: 'normal',  // Allows wrapping of content
+            wordWrap: 'break-word',
+            verticalAlign: vertical_alignment,
         }
     }));
-    const data = args.data;
-
 
     useEffect(() => {
-        const tableHeight = (data.length + 1) * (font_size * 1.5 + border_width + 2 * 8) + 2 * border_width
-        Streamlit.setFrameHeight(tableHeight);
-    }, [data, border_width, font_size]);
+        if (tableRef.current) {
+            const headerHeight = tableRef.current.querySelector('thead')?.offsetHeight || 0;
+            const paginationHeight = paginated ? pagination_extra_table_height : 0;
+
+            const rowHeights = Array.from(
+                tableRef.current.querySelectorAll("tbody tr")
+            ).map(row => (row as HTMLTableRowElement).offsetHeight);
+
+            const totalRowHeight = rowHeights
+                .slice(0, pagination_size_per_page)
+                .reduce((acc: number, height: number) => acc + height, 0);
+
+            const totalHeight = totalRowHeight + headerHeight + paginationHeight + padding;
+            Streamlit.setFrameHeight(totalHeight);
+        }
+    }, [data, pagination_size_per_page, paginated]);
+
+    const rowStyle = (_: any, rowIndex: number): React.CSSProperties => ({
+        backgroundColor: rowIndex % 2 === 0 ? data_bg_color : "",
+        textAlign: data_align,
+    });
+
+    const rowEvents: RowEventHandlerProps<any> = {
+        onMouseEnter: (e, row, rowIndex) => {
+            if (rowIndex !== undefined) {
+                const target = e.currentTarget as HTMLTableRowElement;
+                target.style.backgroundColor = "#f0f0f0"; // Optional: example on-hover styling
+            }
+        }
+    };
 
     return (
-        <div style={table_width ? {width: table_width} : {}}>
+        <div ref={tableRef} style={table_width ? {width: table_width} : {}}>
             <style>{`
+            .custom-border.react-bootstrap-table,
+                .custom-border.react-bootstrap-table table {
+                    border: ${border_width}px ${border_color} solid !important;
+                    margin-bottom: 0 !important;
+                }
+                .custom-border.react-bootstrap-table th,
+                .custom-border.react-bootstrap-table td {
+                    border: ${border_width}px ${border_color} solid !important;
+                }
+                .custom-font-family.react-bootstrap-table td,
+                .custom-font-family.react-bootstrap-table th {
+                    font-family: ${font} !important;
+                    font-size: ${font_size}px !important;
+                }
+                .pagination {
+                     margin-top: 5px;
+                     line-height: ${pagination_line_height}px !important;
+                     justify-content: left !important;
+                     --bs-pagination-color: ${pagination_text_color} !important; 
+                     --bs-pagination-bg: ${pagination_bg_color} !important; 
+                     --bs-pagination-border-color: ${pagination_border_color} !important;
+                     --bs-pagination-border-radius: 0 !important;
+                }
+                .pagination .page-item.active .page-link {
+                    background-color: ${pagination_active_bg_color} !important;
+                    border-color: ${pagination_active_border_color} !important;
+                    color: ${pagination_active_color} !important;
+                }
+                .pagination .page-item .page-link {
+                    color: ${pagination_text_color} !important;
+                     box-shadow: none !important;                     
+                     outline: none !important;                     
+
+                }
+                .pagination .page-item .page-link:hover {
+                    background-color: ${pagination_hover_bg_color} !important;
+                    color: ${pagination_hover_color} !important;
+                    border-color: ${pagination_active_border_color} !important;
+                }
                 .custom-border.react-bootstrap-table,
                 .custom-border.react-bootstrap-table table {
-                border: ${border_width}px ${border_color} solid !important;
-            }
-            .custom-border.react-bootstrap-table th,
-            .custom-border.react-bootstrap-table td {
-            border: ${border_width}px ${border_color} solid !important;
-            }
-            .custom-font-family.react-bootstrap-table td,
-            .custom-font-family.react-bootstrap-table th {
-                font-family: ${font} !important;
-                font-size: ${font_size}px !important;
-        }
-        `}</style>
-
+                    border: ${bordered ? '1px solid' : '0'} ${border_color} !important;
+                }
+                .custom-font-family.react-bootstrap-table td,
+                .custom-font-family.react-bootstrap-table th {
+                    font-family: ${font} !important;
+                    font-size: ${font_size}px !important;
+                }
+                .pagination {
+                    margin-top: 5px;
+                    line-height: ${pagination_line_height}px !important;
+                }
+            `}</style>
             <BootstrapTable
-                keyField='id'
+                keyField="id"
                 data={data}
                 columns={columns}
                 wrapperClasses="custom-border"
                 bordered={bordered}
+                rowStyle={rowStyle}
+                rowEvents={rowEvents}
+                pagination={paginated ? paginationFactory({
+                    page: 1,
+                    sizePerPage: pagination_size_per_page,
+                    paginationSize: pagination_bar_size,
+                    hideSizePerPage: true,
+                }) : undefined}
             />
         </div>
     );