+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2024-10-30
+
+### Added
+
+- vertical alignment
+
+### Changed
+
+- table height calculation for wrapped rows
+
+## [0.0.5] - 2024-10-26
+
+### Added
+
+- configurable pagination
+
 ## [0.0.4] - 2024-10-25
 
 ### Added
@@ -12,15 +28,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - font and font size
 
 ## [0.0.3] - 2024-10-25
-
-### Changed
-
+## Changed
 - removed dependency to bootstrap.min.css
 
 ## [0.0.2] - 2024-10-25
-
-### Fixed
-
+## Fixed
 - build
 
 ## [0.0.1] - 2024-10-25 - deleted